Refactor CMIP6 modules: consolidate models and records; remove legacy cmip6_file_record.py; update imports; lint+format
