+++ src/piddiplatsch/models/__init__.py
@@ -1,3 +1,5 @@
 from .base import HostingNode as HostingNode
-from .cmip6_dataset_model import CMIP6DatasetModel as CMIP6DatasetModel
-from .cmip6_file_model import CMIP6FileModel as CMIP6FileModel
+from .cmip6_model import CMIP6DatasetModel as CMIP6DatasetModel
+from .cmip6_model import CMIP6FileModel as CMIP6FileModel
+
+__all__ = ["CMIP6DatasetModel", "CMIP6FileModel", "HostingNode"]
