+++ tests/test_cmip6_file_record.py
@@ -1,5 +1,5 @@
 from piddiplatsch.records import CMIP6FileRecord
-from piddiplatsch.records.cmip6_file_record import extract_asset_records
+from piddiplatsch.records.cmip6_record import extract_asset_records
 from piddiplatsch.utils.models import parse_pid
 
 # Minimal mock STAC item
